Implement decomposed RSCED and Cplex-based Benders'
Augments the Cplex solver to take advantage of user-provided annotations for Benders' decomposition. Evaluates the variable parameter "_problem" to select the appropriate Benders' strategy. Parameter defaults to "-1" which corresponds to no problem specified. Also, contains an implementation of a decomposable variant to the RSCED formulation.

Note, the current build does not work with Cplex due to problems with lazy constraint generation.